webapi/tct-video-html5-tests/video/video_seek_webm_vp8_vorbis.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this lis
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work withou
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
         Zhang, Zhiqiang <zhiqiang.zhang@intel.com>
         Cao, Jun <junx.cao@intel.com>

-->

<html>
  <head>
    <title>Video Test: video_seek_webm_vp8_vorbis</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-html5-20121025/the-video-element.html#the-video-element" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if user agent is able to seek video playback with currentTime attribute for stream webm vp8 vorbis" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <video id="v">
        Your browser does not support media elements.
    </video>
    <script>
        var t = async_test("video.currentTime - to verify if user agent is able to seek video playback with currentTime attribute for stream webm vp8 vorbis", {timeout:30000});
        setup({timeout:1000000});

        var v = document.getElementById("v");

        var completed = false;

        function startTest() {
            if (completed)
                return false;

            v.currentTime= v.duration - 2;

            t.step(function() {
                assert_true(true, "loadmetadata event");
                assert_true(v.seeking, "seeking flag on seek start should be true");
                v.play();
            });
            return false;
        }

        function seekStarted() {
            if (completed)
                return false;

            t.step(function() {
                assert_true(true, "seeking event");
                v.pause();
            });
            return false;
        }

        function seekEnded() {
            if (completed)
                return false;

            t.step(function() {
                assert_true(true, "seeked event");
                assert_false(v.seeking, "seeking flag on seeked should be false");
                v.play();
            });
            return false;
         }

        function playbackEnded() {
            if (completed)
               return false;

            completed = true;
            t.step(function() {
                assert_true(true, "ended event");
            });
            t.done();
            return false;
        }

        v.addEventListener("ended", playbackEnded, false);
        v.addEventListener("loadedmetadata", startTest, false);
        v.addEventListener("seeking", seekStarted, false);
        v.addEventListener("seeked", seekEnded, false);

        v.src="media/movie_5.webm";
    </script>
  </body>
</html>
